Topi: Add support to detect holes on the underlying surface

+++ src/entities/brick.cpp
@@ -9,11 +9,11 @@ Brick::Brick(EntityIdentificator _id, EntityType _type, SurfaceType surface_type
         IEntity(_id, _type, surface_type, max_states, is_breakable, is_traversable) {
 }
 
-uint16_t Brick::Width() {
+int Brick::Width() {
         return currentSprite.width;
 }
 
-uint16_t Brick::Height() {
+int Brick::Height() {
         return currentSprite.height;
 }
