+++ src/entities/brick.h
@@ -31,8 +31,8 @@ public:
   Brick();
   ~Brick();
   virtual void InitWithSpriteSheet(EntitySpriteSheet*);
-  uint16_t Width();
-  uint16_t Height();
+  int Width();
+  int Height();
   virtual void PrintName();
   bool IsCloud() override;
   bool Update(uint8_t);
