+++ src/entities/cloud.cpp
@@ -9,11 +9,11 @@ Cloud::Cloud(EntityIdentificator _id, EntityType _type, SurfaceType surface_type
         IEntity(_id, _type, surface_type, max_states, is_breakable, is_traversable) {
 }
 
-uint16_t Cloud::Width() {
+int Cloud::Width() {
         return currentSprite.width;
 }
 
-uint16_t Cloud::Height() {
+int Cloud::Height() {
         return currentSprite.height;
 }
