+++ src/entities/side_wall.h
@@ -17,8 +17,8 @@ public:
   SideWall();
   ~SideWall();
   virtual void InitWithSpriteSheet(EntitySpriteSheet*);
-  uint16_t Width();
-  uint16_t Height();
+  int Width();
+  int Height();
   virtual void PrintName();
   bool IsCloud() override;
   bool Update(uint8_t);
