+++ src/sprite.h
@@ -14,10 +14,10 @@ class Sprite
 {
 
 public:
-  uint16_t width;
-  uint16_t height;
-  int16_t xOffset;
-  int16_t yOffset;
+  int width;
+  int height;
+  int xOffset;
+  int yOffset;
   float u1, v1, u2, v2;
   SpriteAreas *areas; // DEPRECATED
   Sprite();
